backend/pyproject.toml
[project]
name = "mail-orchestrator-backend"
version = "0.1.0"
description = "FastAPI backend for Mail Orchestrator"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.0",
  "alembic>=1.13.0",
  "google-auth>=2.0.0",
  "google-auth-oauthlib>=1.2.0",
  "google-api-python-client>=2.0.0",
  "python-dotenv>=1.0.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.setuptools.package-data]
app = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
